Added CFGTransduction task and xor configs
Created new CFGTransduceTask (in tasks/cfgtransduce.py) which gives as input odd positions in a string (or even if you start counting from 0) and requires as output the even (or odd, using python indexing) positions in the string. This allows us to use the grammars we have for prediction to do transduction tasks.

Added configs to use xor grammars (with some restrictions on string lengths)